third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-cascade-multiple-rules.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
div {
    background-color: red;
}
.e {
    --foo: {
        background-color: green;
        width: 10px;
        height: 20px;
    };
    @apply --foo;
    height: 200px;
}
#e {
    width: 100px;
}
</style>
<div class=e id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).width, "100px");
    assert_equals(getComputedStyle(e).height, "200px");
}, "Declarations from @apply cascade correctly across multiple rules");
</script>
